Change the long for int

+++ RevStay/src/main/java/com/Rev/RevStay/controllers/BookingController.java
@@ -34,7 +34,7 @@ public class BookingController {
         return ResponseEntity.ok(bookingResponse);
     }
     @GetMapping("/user/{userId}")
-    public ResponseEntity<List<Booking>> getBookingsByUser(@PathVariable Long userId) {
+    public ResponseEntity<List<Booking>> getBookingsByUser(@PathVariable int userId) {
         List<Booking> bookings = bookingService.getBookingsByUser(userId);
 
         //Check in the bookings if the checkOut pass, set the status for booking as completed
@@ -44,7 +44,7 @@ public class BookingController {
             if (booking.getCheckOut() != null && booking.getCheckOut().isBefore(today)){
                 //The checkOut date pass
                 if (booking.getStatus() != BookingStatus.COMPLETED){
-                    bookingService.updateBookingStatus(booking.getBookId(), BookingStatus.COMPLETED, userId.intValue());
+                    bookingService.updateBookingStatus(booking.getBookId(), BookingStatus.COMPLETED, userId);
                 }
             }
         }
@@ -54,7 +54,7 @@ public class BookingController {
 
     //update booking status
     @PutMapping("/{bookingId}/status")
-    public ResponseEntity<Booking> updateBookingStatus(@PathVariable Long bookingId, @RequestParam BookingStatus bookingStatus, HttpSession session) {
+    public ResponseEntity<Booking> updateBookingStatus(@PathVariable int bookingId, @RequestParam BookingStatus bookingStatus, HttpSession session) {
 
         Booking updatedBooking = bookingService.updateBookingStatus(bookingId, bookingStatus, (Integer) session.getAttribute("userId"));
         return ResponseEntity.ok(updatedBooking);
